requirements.txt
alabaster==0.7.12
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.9b0
bleach==4.1.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
fillpdf==0.1.0
idna==3.2
imagesize==1.2.0
ipykernel==6.3.1
ipyparallel==6.3.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.3
numpy==1.21.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pdf2image==1.16.0
pdfrw==0.4
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
pymssql==2.2.2
pyodbc==4.0.32
pyparsing==2.4.7
PyPDF2==1.26.0
pyreadline==2.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
regex==2021.8.28
requests==2.26.0
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
terminado==0.11.1
testpath==0.5.0
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
TwoCaptcha==0.0.1
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wkhtmltopdf==0.2